api/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
scikit-learn
joblib
